Lesson 9
Multithreading, but did not get any speed up on my 8 core cpu. At first I assumed it was due to the RNG not being thread_local, but doing that actually slowed it down from ~65ms to ~110ms

I'm not quite sure what the issue could be. When running in multithread mode only one core appears to be fully utilized